Add files via upload
Developed a visually appealing and responsive website template for a restaurant specializing in healthy food. This template offers a modern design, intuitive navigation, and an engaging user interface, making it easy for customers to explore the menu, make reservations, and learn more about the restaurant. Built using advanced front-end technologies, this project highlights proficiency in web development and design.